source code and folders necessary for full functionality
Source Code of the Project can be found in the Short_Text_Classification_using_transformers.ipynb
the Cl0 and CL1 folder are used to train the model and AL_Ghazaly folder contains the text that has been tested.
res0M and Model files will contain the results the the saved model respectively.